refactor(nika-dap): the trace forensics come home — the cli keeps the render
The W0 descent (architecture review v2 §1): nika-cli stood at
14,966/15,000 prod LOC (99.8%) with open PRs blocked at the wall. The
forensic half of the trace family descends into nika-dap — the plane
that already holds chain · recover · source_id · stats:

  otel       — the OTLP/JSON projection (takes the chain Verdict
               directly; the exporter version is injected — env! of
               the version stays a CLI concern)
  reproduce  — the comparison taxonomy (Verdict · Row · Report ·
               compare · workflow_of)
  store      — TRACE_DIR · TraceState · TraceMeta · scan · fold_facts
  retention  — RetentionConfig (the three knobs incl. from_env) ·
               Reason · plan · newest_per_workflow · collect → GcReport

The render stays cli-side — the trace_verify shim pattern throughout:
export/reproduce file plumbing + operator messages, the report/line
renderers, the fmt_age/fmt_bytes display vocabulary, the nika-run GC
hook; trace/mod.rs + trace manage untouched (pure render). Every
trace::store / trace::retention consumer reads unchanged via
re-exports at the old paths — zero call-site churn beyond two honest
wildcard arms forced by TraceState going non_exhaustive (FCI-002/016
+ invariant #19 new() on every descended pub struct).

Measured: nika-cli 14,966 → 13,941 prod LOC (92.9% of cap) · nika-dap
1,277 → 2,535. Deps stay L0-only (the absorption is L4-legal). One
unit, two members (D-2026-07-09-N1) — parentage named in
docs/crate-specs/nika-dap.md.

Gates: fmt clean · clippy workspace deny · test workspace lib green ·
check-crate-size OK · run-ci-ratchets 8/8 · the cli integration
batteries incl. trace_export_e2e (the CLI surfaces stay
byte-identical). Clears the wall for the open cli PRs (#352 · #353).
