+++ crates/nika-cli/Cargo.toml
@@ -43,7 +43,7 @@ nika-http        = { path = "../nika-http", version = "0.98.0" }          # Reqw
 nika-exec-runner = { path = "../nika-exec-runner", version = "0.98.0" }   # TokioShell — the exec verb's subprocess runner
 nika-catalog     = { path = "../nika-catalog", version = "0.98.0", features = ["providers", "pricing"] }  # env_var ladder + the rates the preflight shows
 nika-lsp         = { path = "../nika-lsp", version = "0.98.0" }           # `nika lsp` — the language server (stdio · drives the editor extension)
-nika-dap         = { path = "../nika-dap", version = "0.98.0" }           # `nika dap` — the replay debugger (stdio · split from verbs/dap 2026-07-09 · crate-size cap)
+nika-dap         = { path = "../nika-dap", version = "0.98.0" }           # `nika dap` + the trace-forensics plane (chain · recover · otel · reproduce · store · retention — re-exported at the old verb paths)
 nika-mcp         = { path = "../nika-mcp", version = "0.98.0" }           # `nika mcp` — the MCP server (stdio · exposes check/explain to Cursor/Claude Desktop)
 tokio            = { workspace = true }                                   # the bin's async entry (block the run on the executor)
