+++ crates/nika-dap/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name          = "nika-dap"
-description   = "The Debug Adapter Protocol server (`nika dap`) — a READ-ONLY replay debugger over recorded run journals: breakpoints on task lines, stepping walks task settles, stepBack is free because the log is total (replay = re-render, never re-execute). Layer L4 — split from nika-cli's verbs/dap 2026-07-09 (the crate-size cap · the nika-cap precedent)."
+description   = "The Debug Adapter Protocol server (`nika dap`) — a READ-ONLY replay debugger over recorded run journals: breakpoints on task lines, stepping walks task settles, stepBack is free because the log is total (replay = re-render, never re-execute). Also the trace-forensics plane every reader shares: the chain walk, the tolerant recover, the OTLP projection, the reproduce comparison, the store scan and the retention policy. Layer L4 — split from nika-cli's verbs/dap 2026-07-09 (the crate-size cap · the nika-cap precedent), grown by the W0 trace descent the same day (compute descends, render stays)."
 version.workspace      = true
 edition.workspace      = true
 license.workspace      = true
